Remove EmrVisitService in favor of DiagnosisService as all methods are diagnosis related
