+++ api/src/main/java/org/openmrs/module/emrapi/diagnosis/CoreDiagnosisService.java
@@ -3,8 +3,10 @@ package org.openmrs.module.emrapi.diagnosis;
 import org.openmrs.Encounter;
 import org.openmrs.Obs;
 import org.openmrs.Patient;
+import org.openmrs.Visit;
 import org.openmrs.api.context.Context;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -12,6 +14,12 @@ import java.util.List;
  * class implementing the DiagnosisService while delegating calls to the core module
  * */
 public class CoreDiagnosisService {
+
+	private EmrDiagnosisDAO emrDiagnosisDAO;
+
+	public void setEmrDiagnosisDAO(EmrDiagnosisDAO emrDiagnosisDAO) {
+		this.emrDiagnosisDAO = emrDiagnosisDAO;
+	}
 
 	public List<Diagnosis> getDiagnoses(Patient patient, Date fromDate) {
 		return DiagnosisUtils.convert(Context.getDiagnosisService().getDiagnoses(patient, fromDate));
@@ -31,5 +39,14 @@ public class CoreDiagnosisService {
 
 	public List<Obs> codeNonCodedDiagnosis(Obs nonCodedObs, List<Diagnosis> diagnoses){
 		return null;
+	}
+
+	public List<Obs> getDiagnosesAsObs(Visit visit, DiagnosisMetadata diagnosisMetadata, Boolean primaryOnly, Boolean confirmedOnly) {
+		List<org.openmrs.Diagnosis> diagnoses = emrDiagnosisDAO.getDiagnoses(visit, primaryOnly, confirmedOnly);
+		List<Obs> diagnosisList = new ArrayList<Obs>();
+		for (Diagnosis diagnosis : DiagnosisUtils.convert(diagnoses)) {
+			diagnosisList.add(diagnosisMetadata.buildDiagnosisObsGroup(diagnosis));
+		}
+		return diagnosisList;
 	}
 }
